# Optimize RLE encoding