--- a/parquet_purego.go
+++ b/parquet_purego.go
@@ -1,0 +1,5 @@
+//go:build purego || !amd64
+
+package parquet
+
+func optimize(int) bool { return false }