This PR optimizes the RLE encoding, which is used when writing dictionary indexes, repetition and definition levels, and boolean columns.

The encoding of repetition and definition levels are employed when writing repeated or optional columns, in that regard those optimizations participate in completing the work to improve the throughput of writing optional columns (see https://github.com/segmentio/parquet-go/pull/205).

The bit packing for 32 bits integer values was adapted from the algorithm we have in the DELTA_BINARY_PACKED encoding (see https://github.com/segmentio/parquet-go/pull/200).

```
name                old time/op    new time/op     delta
Encode/RLE/boolean     994ns ± 0%      992ns ± 0%    -0.19%  (p=0.001 n=9+9)
Encode/RLE/levels     11.0µs ± 0%      1.9µs ± 0%   -82.44%  (p=0.000 n=9+9)
Encode/RLE/int32      34.8µs ± 0%      3.5µs ± 0%   -89.93%  (p=0.000 n=10+9)

name                old speed      new speed       delta
Encode/RLE/boolean  1.26GB/s ± 0%   1.26GB/s ± 0%    +0.19%  (p=0.000 n=9+9)
Encode/RLE/levels    908MB/s ± 0%   5172MB/s ± 0%  +469.53%  (p=0.000 n=9+9)
Encode/RLE/int32    1.15GB/s ± 0%  11.40GB/s ± 0%  +892.67%  (p=0.000 n=10+9)

name                old value/s    new value/s     delta
Encode/RLE/boolean     10.1G ± 0%      10.1G ± 0%    +0.19%  (p=0.000 n=9+9)
Encode/RLE/levels       908M ± 0%      5172M ± 0%  +469.53%  (p=0.000 n=9+9)
Encode/RLE/int32        287M ± 0%      2849M ± 0%  +892.67%  (p=0.000 n=10+9)
```

I find it interesting to see how byte throughput isn't necessarily what we should be optimizing for. Often times increasing byte throughput requires introducing a lot of complexity in the SIMD algorithms. The encoded value throughput is a lot more important, and tight packing of values in memory is very effective to improve it. The encoding if bit-packed boolean values runs at 10 billion values/s but has the lowest byte throughput, while the encoding of int32 values runs at _only_ 2.8 billion values/s despite having the highest byte throughput (because we read a lot more memory).

Let me know if you have any feedback.